refactor: reuse product card and change loading shimmer color
