+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/component/productCard/ProductCard.kt
@@ -14,6 +14,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
@@ -23,14 +24,19 @@ import mena.dukan_presentation.generated.resources.product_image
 import mena.dukan_presentation.generated.resources.silver_tc
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ProductUiState
+import net.thechance.mena.dukan.presentation.screen.dukanDetails.components.VisiableQuantityComponent
+import net.thechance.mena.dukan.presentation.viewModel.dukanDetails.DukanDetailsUiState
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 
 @Composable
 fun ProductCard(
-    productUiState: ProductUiState,
+    productName: String,
+    productImageUrl: String,
+    productDescription: String,
+    productPrice: Double,
+    productCardBackground: Color? = null,
     productAction: @Composable () -> Unit,
     modifier: Modifier = Modifier
 ) {
@@ -38,7 +44,7 @@ fun ProductCard(
         modifier = modifier
             .fillMaxWidth()
             .background(
-                color = Theme.colorScheme.background.surfaceLow,
+                color = productCardBackground?: Theme.colorScheme.background.surface,
                 shape = RoundedCornerShape(size = Theme.radius.md)
             ).height(104.dp)
             .padding(Theme.spacing._4),
@@ -53,7 +59,7 @@ fun ProductCard(
             )
         ) {
             AsyncImage(
-                model = productUiState.imageUrl,
+                model = productImageUrl,
                 contentDescription = stringResource(Res.string.product_image),
                 modifier = Modifier
                     .size(96.dp)
@@ -71,8 +77,8 @@ fun ProductCard(
                 ),
         ) {
             ProductInfo(
-                name = productUiState.name,
-                description = productUiState.description
+                name = productName,
+                description = productDescription
             )
             Spacer(modifier = Modifier.weight(1f))
             Row(
@@ -82,7 +88,7 @@ fun ProductCard(
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 PriceWithIcon(
-                    price = productUiState.price.toString(),
+                    price = productPrice.toString(),
                     iconRes = Res.drawable.silver_tc,
                     contentDescription = stringResource(Res.string.koin_icon),
                 )
@@ -99,14 +105,29 @@ fun ProductCard(
 private fun ProductCardPreview() {
     MenaTheme {
         ProductCard(
-            ProductUiState(
-                id = "1",
-                imageUrl = "https://calvinklein.scene7.com/is/image/CalvinKlein/LX001376_100_alternate1?wid=1728&qlt=80%2C0&resMode=sharp2&op_usm=0.9%2C1.0%2C8%2C0&iccEmbed=0&fmt=webp",
-                name = "Girls Crochet Tank Top",
-                description = "Girls Crochet Tank Top description text here for this product",
-                price = 39.5
-            ),
-            productAction = { EditProductIcon(onClick = {}) },
+            productName = "Girls Crochet Tank Top",
+            productImageUrl = "https://calvinklein.scene7.com/is/image/CalvinKlein/LX001376_100_alternate1?wid=1728&qlt=80%2C0&resMode=sharp2&op_usm=0.9%2C1.0%2C8%2C0&iccEmbed=0&fmt=webp",
+            productDescription = "Girls Crochet Tank Top description text here for this product",
+            productPrice = 39.5,
+            productCardBackground = Theme.colorScheme.background.surfaceLow,
+            productAction = {EditProductIcon(onClick = {}) },
+            modifier = Modifier.padding(Theme.spacing._12),
+        )
+    }
+}
+
+
+@Preview
+@Composable
+private fun ProductCardPreview2() {
+    MenaTheme {
+        ProductCard(
+            productName = "Girls Crochet Tank Top",
+            productImageUrl = "https://calvinklein.scene7.com/is/image/CalvinKlein/LX001376_100_alternate1?wid=1728&qlt=80%2C0&resMode=sharp2&op_usm=0.9%2C1.0%2C8%2C0&iccEmbed=0&fmt=webp",
+            productDescription = "Girls Crochet Tank Top description text here for this product",
+            productPrice = 39.5,
+            productCardBackground = Theme.colorScheme.background.surfaceLow,
+            productAction = {VisiableQuantityComponent(DukanDetailsUiState())  },
             modifier = Modifier.padding(Theme.spacing._12),
         )
     }
