+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/shelfDetails/ShelfDetailsScreen.kt
@@ -15,7 +15,7 @@ import net.thechance.mena.designsystem.presentation.component.scaffold.Scaffold
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
 import net.thechance.mena.dukan.presentation.navigation.LocalNavController
-import net.thechance.mena.dukan.presentation.screen.dukanDetails.components.ManageShelfProducts
+import net.thechance.mena.dukan.presentation.screen.dukanDetails.components.ShelfProducts
 import net.thechance.mena.dukan.presentation.util.ObserveAsEffect
 import net.thechance.mena.dukan.presentation.util.OnSystemBackPressed
 import net.thechance.mena.dukan.presentation.util.pagination.Pager
@@ -94,7 +94,7 @@ private fun ShelfDetailsContent(
             )
         }
     ) {
-        ManageShelfProducts(
+        ShelfProducts(
             state = state,
             pager = pager,
         )
