+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/viewModel/dukanDetails/DukanDetailsUiState.kt
@@ -10,6 +10,7 @@ data class DukanDetailsUiState(
     val shelvesState: ShelvesState = ShelvesState.LOADING,
     val productsShelf: PagingData<ProductUiState> = PagingData(),
     val productsState: ProductsState = ProductsState.LOADING,
+    val showProductQuantity: Boolean = false,
     val shelfIdSelected: String? = null
 ) {
     data class DukanInfo(
@@ -49,7 +50,6 @@ data class DukanDetailsUiState(
         val imageUrl: String = "",
         val price: Double = 0.0,
         val description: String = "",
-        val showProductQuantity: Boolean = false
     )
 
     enum class ProductsState {
